refactor: no more detail namespaces
