--- include/blackhole/detail/spinlock.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#ifdef __linux__
-#   include "blackhole/detail/spinlock.linux.hpp"
-#elif __APPLE__
-#   include "blackhole/detail/spinlock.osx.hpp"
-#endif
