+++ src/config/factory.cpp
@@ -1,7 +1,8 @@
 #include "blackhole/config/factory.hpp"
 
 #include "blackhole/forward.hpp"
-#include "blackhole/detail/util/deleter.hpp"
+
+#include "../util/deleter.hpp"
 
 namespace blackhole {
 inline namespace v1 {
@@ -9,9 +10,9 @@ namespace config {
 
 factory_t::~factory_t() = default;
 
-}  // namespace config
+} // namespace config
 
 template auto deleter_t::operator()(config::factory_t* value) -> void;
 
-}  // namespace v1
-}  // namespace blackhole
+} // namespace v1
+} // namespace blackhole
