+++ src/formatter/tskv.cpp
@@ -12,11 +12,10 @@
 #include "blackhole/formatter.hpp"
 #include "blackhole/record.hpp"
 
-#include "blackhole/detail/memory.hpp"
-#include "blackhole/detail/util/deleter.hpp"
-
 #include "../attribute.hpp"
 #include "../datetime.hpp"
+#include "../memory.hpp"
+#include "../util/deleter.hpp"
 #include "string/token.hpp"
 
 namespace blackhole {
