+++ src/sink/syslog.hpp
