+++ src/spinlock.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#ifdef __linux__
+#   include "spinlock.linux.hpp"
+#elif __APPLE__
+#   include "spinlock.osx.hpp"
+#endif
