+++ src/spinlock.osx.hpp
@@ -4,15 +4,12 @@
 
 namespace blackhole {
 inline namespace v1 {
-namespace detail {
 
 class spinlock_t {
     os_unfair_lock mutex;
 
 public:
-    constexpr spinlock_t() noexcept:
-        mutex({0})
-    {}
+    constexpr spinlock_t() noexcept : mutex({0}) {}
 
     auto lock() noexcept -> void {
         os_unfair_lock_lock(&mutex);
@@ -27,6 +24,5 @@ public:
     }
 };
 
-}  // namespace detail
-}  // namespace v1
-}  // namespace blackhole
+} // namespace v1
+} // namespace blackhole
